modifying pardiso

+++ src/system.hh
@@ -3,5 +3,5 @@
 #include <vector>
 #include "mkl.h"
 
-void formRHS(std::vector<float> &vec, MKL_INT size);
+void formRHS(std::vector<double> &vec, MKL_INT size);
 void formA(sparse_matrix_t &A, MKL_INT size);
